refactor(builder): split builder.rs into modular per-element structure
Split the 452-line builder.rs into a modular directory structure with
dedicated modules for each AST element type:

- builder/mod.rs: Main dispatcher and PatternMatch enum (121 lines)
- builder/builders/: Element-specific builder modules
  - helpers.rs: Shared utility functions
  - verbatim.rs: Verbatim block builder
  - annotation.rs: Annotation builders (block & single-line)
  - list.rs: List builder
  - definition.rs: Definition builder
  - session.rs: Session builder
  - paragraph.rs: Paragraph builder

Benefits:
- Improved code navigation and maintainability
- Each element type in its own focused module
- Easier to modify individual builders independently
- Consistent with modular patterns used elsewhere in the codebase
- Reduced main module from 452 to 121 lines

All tests passing (253 lib tests + 13 doc tests + 45 integration tests).
