+++ lex-parser/src/lex/parsing/parser/builder/builders/mod.rs
@@ -1,0 +1,22 @@
+//! AST Node Builders
+//!
+//! This module contains builder functions for converting matched grammar patterns
+//! into ParseNode AST structures. Each element type has its own dedicated builder module.
+
+mod annotation;
+mod definition;
+mod helpers;
+mod list;
+mod paragraph;
+mod session;
+mod verbatim;
+
+pub(in crate::lex::parsing::parser::builder) use annotation::{
+    build_annotation_block, build_annotation_single,
+};
+pub(in crate::lex::parsing::parser::builder) use definition::build_definition;
+pub(in crate::lex::parsing::parser::builder) use list::build_list;
+pub(in crate::lex::parsing::parser::builder) use paragraph::build_paragraph;
+pub(in crate::lex::parsing::parser::builder) use session::build_session;
+pub(in crate::lex::parsing::parser::builder) use verbatim::build_verbatim_block;
+pub(in crate::lex::parsing::parser) use verbatim::VerbatimGroupMatch;
